Add files via upload
In this half-space cooling for ocean depth modelling, I use the information from heat flow to model the cooling of the oceanic lithosphere and derive predicted depths for the ocean basins. Datasets used in the analysis were obtained from Smith & Sandwell (1997) for the bathymetry and Müller et al. (2018) for the crustal age map.